chore: Update Docker Compose files and Dockerfile for backend and frontend services
