+++ backend/app/config.py
@@ -1,0 +1,5 @@
+from starlette.config import Config
+
+config = Config(".env")
+
+CORS_ORIGINS = config("CORS_ORIGINS", cast=lambda v: [s.strip() for s in v.split(",")])
