+++ backend/app/routes/health.py
@@ -1,0 +1,13 @@
+from fastapi import APIRouter
+from starlette.responses import JSONResponse
+from ..database import cursor
+
+router = APIRouter()
+
+@router.get("/health")
+async def health_check():
+    try:
+        cursor.execute('SELECT 1')
+        return JSONResponse(content={"status": "healthy"})
+    except Exception as e:
+        return JSONResponse(content={"status": "unhealthy", "error": str(e)}, status_code=500)
